Move all the badges so they're all together!
Get Your Shit Together. Get it all together and put it in a backpack, all your shit, so it's together. And if you got to take it somewhere, take it somewhere you know, take it to the shit store and sell it, or put it in a shit museum, I don't care what you do. Just gotta get it together. Get your shit together.